Move FragDetails to ActDetails
usage of a separate Fragment for the details screen is redundant and adds unnecessary complexity, this is an initial refactor that moves all related display code and loading functionality into ActDetails.

TODO lots of cleanup edits / attach FragImageViewer (move into separate file) / lots of other smaller things
