+++ app/src/main/java/com/derek_s/hubble_gallery/internal/di/CoreGraph.java
@@ -2,7 +2,6 @@ package com.derek_s.hubble_gallery.internal.di;
 
 import com.derek_s.hubble_gallery.base.BaseActivity;
 import com.derek_s.hubble_gallery.base.HubbleApplication;
-import com.derek_s.hubble_gallery.ui.fragments.FragDetails;
 import com.derek_s.hubble_gallery.ui.fragments.FragMain;
 import com.derek_s.hubble_gallery.ui.fragments.FragNavigationDrawer;
 
@@ -16,8 +15,6 @@ public interface CoreGraph {
     void inject(BaseActivity app);
 
     void inject(FragNavigationDrawer app);
-
-    void inject(FragDetails app);
 
     void inject(FragMain app);
 }
